Rework favicons
Added multiple versions of the favicon and a favicon.ico for the search engines and as the standard for the browser.

The same favicon does now work with the dark mode of the browser as well as the favicons now have a white background that is distancing itself from the black background.

This favicon can now be used for mobile devices.

The site.webmanifest file allows the website to be used as a PWA for users that want a quick link to the website and the ones that want is isolated.