Fixed Beacon
Changed it so that the method doesn't throw an error in the loop (EFFECTS_BY_LEVEL goes from 0 to 3). Added a toString in the first if statement because it would always return false and automatically select the first effect when you specify the second one.

Fixed another small bug with beacons

Another small fix